requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
aiostream==0.4.5
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
bert-extractive-summarizer==0.10.1
blis==0.7.9
Bottleneck @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_07078715-3ab7-4562-8d3d-d56b0eaa0f7dp504n_ny/croots/recipe/bottleneck_1657175566567/work
Brotli==1.0.9
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.0.0
commonmark==0.9.1
confection==0.0.3
cymem==2.0.7
easy-ocr==0.1.1
easyocr==1.6.2
fastapi==0.85.1
ffmpeg-python==0.2.0
filelock==3.8.0
Flask==2.2.2
Flask-WTF==1.0.1
frozenlist==1.3.1
future==0.18.2
grpclib==0.4.3
gunicorn==20.1.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.10.1
hyperframe==6.0.1
idna==3.4
imageio==2.4.1
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.1
modal-client @ https://modal.com/api/client-library/us-E1bdf1izjHZJpPOBlA8qUj/modal_client-0.39.125-py3-none-any.whl
more-itertools==9.0.0
multidict==6.0.2
murmurhash==1.0.9
mutagen==1.46.0
networkx==2.8.8
ninja==1.11.1
nltk==3.7
numexpr @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_98sxk8lloi/croot/numexpr_1668713874222/work
numpy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8efwh7lkf0/croot/numpy_and_numpy_base_1668593730126/work
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.5.1
pathy==0.10.0
Pillow==9.0.0
preshed==3.0.8
protobuf==4.21.8
pyclipper==1.3.0.post4
pycryptodomex==3.15.0
pydantic==1.10.2
Pygments==2.13.0
pyparsing @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_3b_3vxnd07/croots/recipe/pyparsing_1661452540919/work
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytube==12.1.0
pytz @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_830e8540-4c59-42fe-9515-a6f79c24ff6dh653xa67/croots/recipe/pytz_1654762630628/work
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
rich==12.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.10.1
Shapely==1.8.5.post1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==5.2.1
sniffio==1.3.0
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
starlette==0.20.4
synchronicity==0.2.10
tblib==1.7.0
tesserocr==2.5.2
textwrap3==0.9.2
thinc==8.1.5
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.13.1
toml==0.10.2
torch==1.12.1
torchvision==0.14.0
tqdm==4.64.1
transformers==4.23.1
typer==0.6.1
types-certifi==2021.10.8.3
types-toml==0.10.8
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
whisper @ git+https://github.com/openai/whisper.git@9f70a352f9f8630ab3aa0d06af5cb9532bd8c21d
WTForms==3.0.1
yarl==1.8.1
yt-dlp==2022.10.4
zipp==3.9.0
